100 fix

+++ 0x1E-search_algorithms/100-jump.c
@@ -14,7 +14,7 @@
 
 int jump_search(int *array, size_t size, int value)
 {
-size_t jump_step, prev = 0;
+unsigned long int jump_step, prev = 0;
 
 if (array == NULL || size == 0)
 return (-1);
